requirements.txt
transformers~=4.6.1
datasets~=1.6.2
tokenizers~=0.10.2
tqdm~=4.49.0
PyGithub~=1.55
GitPython~=3.1.17

-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cu111
torchvision==0.9.1+cu111
torchaudio===0.8.1